ml-service/requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
scikit-learn
joblib
pytest
requests
python-dotenv
matplotlib
seaborn
